% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMARETTOHub_BuildApp.R
\name{AMARETTOHub_BuildApp}
\alias{AMARETTOHub_BuildApp}
\title{AMARETTOHub_BuildApp}
\usage{
AMARETTOHub_BuildApp(AppTitle, con_info, Community_Info)
}
\arguments{
\item{AppTitle}{The title of the App}

\item{con_info}{The list of necessary Neo4j server information: url, username and password}

\item{Community_Info}{A list of information on Community AMARETTO: (1) a vector of cohort names and (2) type of nodes in Community AMARETTO}
}
\description{
Building the AMARETTOHub shiny+rmarkdown app
}
