% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMARETTOHub_InitiateEntities.R
\name{AMARETTOHub_InitiateEntities}
\alias{AMARETTOHub_InitiateEntities}
\title{AMARETTOHub_InitiateEntities}
\usage{
AMARETTOHub_InitiateEntities(
  Neo4j_Dir,
  AMARETTOlink,
  cAMARETTO_Results,
  AMARETTO_Results,
  AMARETTOinits,
  hgtest_tbl_all,
  phenotype_tests_all
)
}
\arguments{
\item{Neo4j_Dir}{the path to directory that Neo4j files to be stored.}

\item{AMARETTOlink}{The link to the HTML reports}

\item{cAMARETTO_Results}{Community AMARETTO result}

\item{AMARETTO_Results}{The list of AMARETTO results}

\item{AMARETTOinits}{The list of AMARETTO init objects}

\item{hgtest_tbl_all}{The list of all Functional Enrichments of Modules tables}

\item{phenotype_tests_all}{The list of all Phenotype association tables}
}
\description{
This function creates necessary CSV files to import Entities to Neo4j server
}
