% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMARETTOHub_InitiateEntities.R
\name{AMARETTOHub_InitiateEntities}
\alias{AMARETTOHub_InitiateEntities}
\title{AMARETTOHub_InitiateEntities}
\usage{
AMARETTOHub_InitiateEntities(
  AMARETTODirectory,
  AMARETTOlink,
  cAMARETTO_Results,
  AMARETTO_Results,
  AMARETTOinits,
  hgtest_tbl_all,
  phenotype_tests_all
)
}
\arguments{
\item{AMARETTODirectory}{The AMARETTO Directory}

\item{AMARETTOlink}{The link to the HTML reports}

\item{cAMARETTO_Results}{Community AMARETTO results}

\item{AMARETTO_Results}{The list of AMARETTO results}

\item{AMARETTOinits}{The list of AMARETTO init objects}

\item{hgtest_tbl_all}{The data table of Module vs Genesets}

\item{phenotype_tests_all}{The data table of Module vs Phenotypes}
}
\description{
This function creates necessary CSV files to import Entities to Neo4j server
}
